services/worldgraph/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
pydantic-settings
httpx
google-cloud-storage
